requirements.txt
boto3==1.9.79
botocore==1.12.79
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
PyJWT==1.7.1
python-dateutil==2.7.5
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
